requirements.txt
Flask
Werkzeug
Markdown
requests
Pillow
vercel-blob
pymdown-extensions
PyYAML
python-dotenv 